pyproject.toml
[tool.poetry]
name = "korean-time"
version = "0.1.0"
description = ""
authors = ["Kevin Y. Chen"]

[tool.poetry.dependencies]
python = "^3.9"
google-cloud-texttospeech = "^2.12.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
